\name{Msa_4x_chr3}
\alias{Msa_4x_chr3}
\docType{data}
\title{
Read depth data for tetraploid \emph{Miscanthus sacchariflorus} on chromosome 3
}
\description{
Allelic read depth data from a RAD-seq experiment across 270 tetraploid
\emph{Miscanthus sacchariflorus} and 3414 loci on chromsome 3.
}
\usage{data("Msa_4x_chr3")}
\format{
  A \code{RADdata} object generated by \code{VCF2RADdata}, prior to genotype calling.
}
\source{
Lindsay V. Clark, Xiaoli Jin, Karen K. Petersen, Kossanou G. Anzoua, Larissa Bagmet, Pavel Chebukin, Martin Deuter, Elena Dzyubenko, Nicolay Dzyubenko, Kweon Heo, Douglas A. Johnson, Uffe Jorgensen, Jens B. Kjeldsen, Hironori Nagano, Junhua Peng, Andrey Sabitov, Toshihiko Yamada, Ji Hye Yoo, Chang Yeon Yu, Stephen P. Long, and Erik J. Sacks (2019) "Population structure of \emph{Miscanthus sacchariflorus} reveals two major polyploidization events, tetraploid-mediated unidirectional introgression from diploid \emph{M. sinensis}, and diversity centered around the Yellow Sea." Annals of Botany 124(4): 731-748. \doi{10.1093/aob/mcy161}.

Data available in VCF format on the Illinois Data Bank: \doi{10.13012/B2IDB-8170405_V1}
}
\references{
Marker positions are with respect to the \emph{M. sinensis} v7 reference genome:
\url{https://phytozome.jgi.doe.gov/pz/portal.html#!info?alias=Org_Msinensis_er}

Therese Mitros, Adam M. Session, Brandon T. James, Guohong Albert Wu, Mohammad B. Belaffif, Lindsay V. Clark, Shengqiang Shu, Hongxu Dong, Adam Barling, Jessica R. Holmes, Jessica E. Mattick, Jessen V. Bredeson, Siyao Liu, Kerrie Farrar, Katarzyna Glowacka, Stanislaw Jezowski, Kerrie Barry, Won Byoung Chae, John A. Juvik, Justin Gifford, Adebosola Oladeinde, Toshihiko Yamada, Jane Grimwood, Nicholas H. Putnam, Jose De Vega, Susanne Barth, Manfred Klaas, Trevor Hodkinson, Laigeng Li, Xiaoli Jin, Junhua Peng, Chang Yeon Yu, Kweon Heo, Ji Hye Yoo, Bimal Kumar Ghimire, Iain S. Donnison, Jeremy Schmutz, Matthew E. Hudson, Erik J. Sacks, Stephen P. Moose, Kankshita Swaminathan, and Daniel S. Rokhsar (2020) "Genome biology of the paleotetraploid perennial biomass crop \emph{Miscanthus}."" Nature Communications 11:5442. \doi{10.1038/s41467-020-18923-6}
}
\examples{
data(Msa_4x_chr3)
Msa_4x_chr3
}
\keyword{datasets}
